Setup Player controller Tick Method Pt2 and OutParameter methods
Pt 1: added .ipch files to git lfs
pt2: pushed playercontroller.h and .cpp

For aiming with player crosshair a method GetSightRayHitLocation() is
created.
Returns a boolean.
an if statement passes an out parameter FVector &OutHitLocation (a
variable that when contents are changed in the method the contents are
also changed outside the method)  into the GetSightRayHitLocation()
method.

If true is returned i.e. if the ray intersects an object in the scene
log the FVector of the OutHitLocation

Unreal Bug: Tick function for PlayerController will fire outside of play
if Blueprint editor is open in the engine. (Make sure to protect your
pointers)