Bug Reporting and Cleaning up dependencies by forward declaring.
Bug Reporting:
1. Show steps to reproduce the bug in code. Comment on the line of the
code with the bug exactly what will cause it to break.
2. Provide snap shot of code
3. Post a link to git hub commit where Unreal can get the bug.

Note About Forward Declarations:
- If you are "requiring" another class in a header file create the
forward declaration in the header file and add the #include for that
class in the .cpp file
(e.g. in TankAIController.h we use ATank* GetControlledTank(). Because
this "requires" ATank class we need to do forward declaration in the
header file and add a #include "Tank.h" in the TankAIController.cpp
file)
-#include typically goes above the class name.h