Using Reference Viewer to track asset dependencies
When deleting assets Unreal will tell you if an asset depends on the
assets you are deleting.

1. Right click Asset>> Select ReferenceViewer - This will show all the
references for that asset

2. Select all the references the asset is dependent>> right click>>
Select Find in content browser>> Then move the assets to your content
folder.

3. Right click The asset folder (StarterContent)>> Select fix up
redirectories in folder.

4. Re-Run the delete

Repeat this for every asset with dependencies and then when there are no
more listed you may delete or force delete