Observer Pattern
Note: In this example Tank is the Broadcaster, Controller is the
Listener

Steps for setting up Dynamic Multicast Delegates:
1. Create Type: DECLARE_DYNAMIC_MULTICAST_DELEGATE(FSubjectName);
- DECLARE_DYNAMIC_MULTICAST_DELEGATE(FTankDelegate);
-this is defined between the #includes and class definition of
Broadcaster (Tank.h)
2. Declare: FSubjectName OnSomethingHappened;
-  FTankDelegate OnDeath;
- This is defined in public section of Broadcaster (Tank.h)
3. Broadcast: OnSomethingHappened.Broadcast();
- OnDeath.Broadcast();
- this is Defined in section where the event occurs for Broadcaster
(Tank.cpp)
- for example during the instance when the tank health goes to 0
broadcast it is dying

---------------------------------------------------------------------------

4. Declare delegate on listener(s) :
- declare a delegate method in listener.h that is a UFUNCTION() that
will handle the delegate when it gets called (TankAIController.h)
- UFUNCTION()
void OnPossessedTankDeath();
- Define it in the listener.cpp file. this will be called when the
listener receives a broadcast that triggers the method to do something

void ATankAIController::OnPossessedTankDeath() {
UE_LOG(LogTemp, Warning, TEXT("RECEIVED"));
}

5: Register to listen in the listener.cpp file:
BroadcastingInstance-> OnSomethingHappened.AddUniqueDynamic(this,
&AListenerClass::DelegateMethod);

-(TankAIController.cpp):  PossessedTank->OnDeath.AddUniqueDynamic(this,
&ATankAIController::OnPossessedTankDeath);

- BroadcastingInstance->SomethingHappened says Find the thing that is
broadcasting
- .AddUniqueDynamic(this, &AListenerClass::DelegateMethod); says tell
the listener (this) to Go find in memory the address of the Delegate
method (&AListenerClass::DelegateMethod)
- The method defined in step 4 is then called and the code in that block
is executed

When Do We Register our listener?
- We will register when the listener (TankAIController) is possessed by
using a virtual method called SetPawn. SetPawn will get called when the
pawn is possessed. The reason we dont do this somewhere else like
BeginPlay is because we cant be sure that the tank is possessed yet when
BeginPlay is executed
- in TankAIController.h:  override the SetPawn() Method
- in TankAIController.cpp: Implement the SetPawn() Method:
- figure out what the possessed tank is: Grab the Pawn and cast it as an
ATank type
auto PossessedTank = Cast<ATank>(InPawn);
- #include Tank.h so that SetPawn is able to cast the InPawn
- Then finally we can subscribe to the local Tank death event method
OnPossessedTankDeath() from step 5
PossessedTank->OnDeath.AddUniqueDynamic(this,
&ATankAIController::OnPossessedTankDeath);